spark/base: [Enh] Move exception base class to details (#32)
