+++ spark/base/include/spark/base/details/Exception.h
@@ -1,5 +1,9 @@
 #pragma once
 
+#include <format>
+#include <source_location>
+#include <stdexcept>
+
 #define SPARK_BASE_DETAILS_THROW_EXCEPTION_2_ARG(ExceptionType, additionalMessage) throw ExceptionType(additionalMessage)
 #define SPARK_BASE_DETAILS_THROW_EXCEPTION_1_ARG(ExceptionType)                    SPARK_BASE_DETAILS_THROW_EXCEPTION_2_ARG(ExceptionType, nullptr)
 #define SPARK_BASE_DETAILS_GET_3_TH_ARG(arg1, arg2, arg3, ...)                     arg3
@@ -23,3 +27,38 @@
  * \endcode
  */
 #define SPARK_BASE_DETAILS_THROW_EXCEPTION_MACRO_CHOOSER(...) SPARK_BASE_DETAILS_GET_3_TH_ARG_IN_SET((__VA_ARGS__, SPARK_BASE_DETAILS_THROW_EXCEPTION_2_ARG, SPARK_BASE_DETAILS_THROW_EXCEPTION_1_ARG)) (__VA_ARGS__)
+
+SPARK_WARNING_PUSH
+SPARK_DISABLE_MSVC_WARNING(4275) // non dll-interface class used as base for dll-interface class -> disabled since class is an stl class
+
+namespace spark::base::details
+{
+    /**
+     * \brief The base class for all exceptions in Spark. Should be used as a base class only.
+     */
+    class SPARK_BASE_EXPORT Exception : public std::runtime_error
+    {
+    public:
+        explicit Exception(const char* class_name, const char* message, const std::source_location& source_location)
+            : std::runtime_error(FormatMessage(class_name, message, source_location).c_str()) {}
+
+        ~Exception() override = default;
+
+        Exception(const Exception& other) = default;
+        Exception(Exception&& other) noexcept = default;
+        Exception& operator=(const Exception& other) = default;
+        Exception& operator=(Exception&& other) noexcept = default;
+
+    protected:
+        static std::string FormatMessage(const char* class_name, const char* message, const std::source_location& source_location)
+        {
+            return std::format("Exception of type {0} has been thrown.\nError happen in {1}:{2}.\nException contains the message:\n{3}",
+                               class_name,
+                               source_location.file_name(),
+                               source_location.line(),
+                               message);
+        }
+    };
+}
+
+SPARK_WARNING_POP
